// apps/console/src/lib/api.ts
export interface GeoLocation {
  ip?: string;
  hostname?: string;
  city?: string;
  region?: string;
  country?: string;
  loc?: string;
  org?: string;
  postal?: string;
  timezone?: string;
}
// ...
export interface Agent {
  id?: string;
  machine_id?: string;
  name?: string;
  os?: string;
  platform?: string;
  kernel_version?: string;
  arch?: string;
  maintenance_mode?: boolean;
  reporting_interval_seconds?: number;
  created_at?: string;
  /** @nullable */
  deleted_at?: string | null;
  last_seen?: string;
  location?: GeoLocation;
  meta?: string;
  /** Present in list response */
  monitor_count?: number;
  /**
   * From latest AgentReport.Location.IP
   * @nullable
   */
  ip?: string | null;
  /**
   * From latest AgentReport
   * @nullable
   */
  uptime_seconds?: number | null;
}

export interface AgentReport {
  id?: string;
  agent_id?: string;
  created_at?: string;
  uptime_seconds?: number;
  timestamp?: string;
  cpu?: CPUStats;
  memory?: MemoryStats;
  disk?: DiskStats;
  location?: GeoLocation;
}

export interface Monitor {
  id?: string;
  /** @nullable */
  description?: string | null;
  type?: string;
  name?: string;
  agent_id?: string;
  /** @nullable */
  last_successful_report_at?: string | null;
  reporting_interval_seconds?: number;
  computed_health?: string;
  /** @nullable */
  last_health_computation?: string | null;
  lifecycle?: string;
  health?: string;
  meta?: string;
  created_at?: string;
  updated_at?: string;
  /** @nullable */
  deleted_at?: string | null;
}

export interface MonitorReport {
  id?: string;
  monitor_id?: string;
  payload?: string;
  collected_at?: string;
  health?: string;
  created_at?: string;
}

export type ListAgentsResponseData = {
  agents?: Agent[];
  count?: number;
  limit?: number;
  offset?: number;
};
// ...
export type ListMonitorsResponseData = {
  monitors?: Monitor[];
  count?: number;
  limit?: number;
  offset?: number;
};